+++ lib/QC_Mutex.cpp
@@ -45,7 +45,7 @@ static AbstractQoreNode *MUTEX_lock(QoreObject *self, SmartMutex *m, const QoreL
 
 // Mutex::lock(timeout $timeout) returns int
 static AbstractQoreNode *MUTEX_lock_to(QoreObject *self, SmartMutex *m, const QoreListNode *params, ExceptionSink *xsink) {
-   int timeout_ms = HARD_QORE_INT(params, 0);
+   int timeout_ms = (int)HARD_QORE_INT(params, 0);
    int rc = m->grab(xsink, timeout_ms);
    return !*xsink ? new QoreBigIntNode(rc) : 0;
 }
